move user type to @/types/user
